requirements.txt
azure-storage-blob
pandas
scikit-learn
joblib